mainpanel: Clarified instructions for editing relation table
Moved the "Page Looks Right" button to halfway through the instructions,
hides the first part of instructions and this button after it is pressed.
New Col button is clarified to be distinct from the color blocks. Column
headers in the relation preview table are color coded to the column color.